requirements.txt
python-dotenv
openai
langchain
langchain_openai
langchainhub
neo4j
youtube-search